+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 2.5 (pre-release):
 * [Fix] Better handle `renderer` parameter in: [`Meshes.RescaleTextureToLength`](https://ihsoft.github.io/KSPDev_Utils/v2.5/html/M_KSPDev_ModelUtils_Meshes_RescaleTextureToLength.htm).
-* [Change] INCOMPATIBLE! Fix `SafeDestory` method name spelling: [`Hierarchy.SafeDestroy`](https://ihsoft.github.io/KSPDev_Utils/v2.5/html/M_KSPDev_ModelUtils_Hierarchy_SafeDestroy.htm).
 * [Change] Dont fail if id/name of the part is not found in `PartNodePatcher.GetPartId`.
+* [Change] INCOMPATIBLE! Fix `SafeDestory` method name spelling: [`Hierarchy.SafeDestroy`](https://ihsoft.github.io/KSPDev_Utils/v2.5/html/M_KSPDev_ModelUtils_Hierarchy_SafeDestroy.htm).
 * [Change] INCOMPATIBLE! Refactoring the `CheckClickEvent` method: [`EventChecker.CheckClickEvent`](https://ihsoft.github.io/KSPDev_Utils/v2.5/html/M_KSPDev_InputUtils_EventChecker_CheckClickEvent.htm).
 * [Change] INCOMPATIBLE! Refactoring the `UpdateColliders` method: [`Colliders.UpdateColliders`](https://ihsoft.github.io/KSPDev_Utils/v2.5/html/M_KSPDev_ModelUtils_Colliders_UpdateColliders.htm).
 * [Change] INCOMPATIBLE! Refactoring the `SimpleStateMachine`: once the machine is shutdown, it cannot be used anymore. [`ProcessingUtils.SimpleStateMachine`](https://ihsoft.github.io/KSPDev_Utils/v2.5/html/T_KSPDev_ProcessingUtils_SimpleStateMachine.htm).
@@ -9,6 +9,7 @@
 * [Enhancement] Add new method: [`Meshes.CreatePrimitiveWithCollider`](http://ihsoft.github.io/KSPDev_Utils/v2.5/html/M_KSPDev_ModelUtils_Meshes_CreatePrimitiveWithCollider.htm).
 * [Enhancement] Add methods for better collidres control: [`Colliders.CreatePrimitiveWithCollider`](http://ihsoft.github.io/KSPDev_Utils/v2.5/html/M_KSPDev_ModelUtils_Meshes_CreatePrimitiveWithCollider.htm).
 * [Enhancement] Add a callback to notify if a part being created by the EVA editor: [`AbstractPartModule.OnEvaPartLoaded`](http://ihsoft.github.io/KSPDev_Utils/v2.5/html/M_KSPDev_PartUtils_AbstractPartModule_OnEvaPartLoaded.htm).
+* [Enhancement] Add parameter `exclude` into highlight updater: [`PartModel.UpdateHighlighters`](http://ihsoft.github.io/KSPDev_Utils/v2.5/html/M_KSPDev_PartUtils_PartModel_UpdateHighlighters.htm).
 
 # 2.4 (July 18th, 2020):
 * [Fix #9] Events and actions don't localize.